+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.32.0] - 2022-22-09
+
+### Added
+
+- Added support for handling the below operators in targeting and whitelisting i.e.
+  * Greater than
+  * Greater than equal to
+  * Less than
+  * Less than equal to
+
 ## [1.31.0] - 2022-25-03
 
 ### Changed